Tweak getusec implementation; fix regressions (?).
There are some cases where tsnow is not yet initialized (before main
is called) so ensure that getusec calls clock_gettime on the first
invocation.

Allow the option of making a syscall on getusec for cases where
we may be tracking latency of RPC processing (that does not happen
to transit async/core.C); without the syscall, all processing appears
to take 0 time.