Tweak BerkeleyDB parameters for improved write performance.
Push the log buffer from the default of 32KB to 4MB so that more
transactions can be buffered.  Then, be even less careful insistent
about writing logs to disk by using TXN_NOSYNC.  Since we checkpoint
approximately every 5m, we can lose up to 5m of transactions (or
however long it takes to fill 4MB of log buffer) during recovery.

Update cache size to 30MB which ought to be enough to buffer up
at least ten seconds of full speed writes on a single host (and
should be more than enough if we have more than one host).  This
may be overkill and also may turn out to be a problem when we run
lots of vnodes on a single machine.

This seems to keep adbd STORE RPC latency down to a median of 10ms.