Special-case replication handling for IDA: don't encode.
Replication is simply IDA erasure coding with dfrags = 1.  However
there is no point in doing a lot of math to encode the entire content
of the object for the different replicas.  This is a simple hack that
checks for dfrags == 1 and quickly encodes the entire raw fragment after
a simple header.