Recursively grow merkle_tree during insert.
If all the keys in a leaf are in a single branch, converting it to an
internal node once will not help; we must create enough additional levels
so that the nodes are each smaller than 64 keys.

Add the incremental test case to the tester to force this kind of behavior.