Run maintenance cycle even when only one neighbor is seen.
As long as there is a predecessor, there is someone to sync with.
So, might as well sync with them, even if it is unlikely that any
productive repairs can get done right now---there may be repairs
needed later when more nodes are in the system.

This also helps make sure that getrepairs will have a merkle_tree_bdb
to open up read-only, regardless of the number of neighbors.