Separate calling adbd's sync from maintenance process.
Set up a separate "thread" that runs as frequently as reptm.
This way we can run without maintenance enabled but not
accumulate a huge collection of log files.

We used to have a separate synctm thread, which I removed back in
2006 when originally introducing the current adbd format (in
changeset af834417978e).  At that time, there was no checkpointing
going on at all (b/c no one was calling sync); probably a mistake.
This year I re-introduced the ability to call sync from outside of
adbd and this change re-introduces the fact that it should not be
linked to repair.  Perhaps one day I'll go back to the idea that
adbd should sync the database automatically (which actually sounds
good to me at the moment but seems like more work right now).  Sigh.