Replace suio_flatten with suio::copyout.
In all of our uses of suio_flatten, we allocate space somewhere,
call suio_flatten, memcpy the results into our pre-allocated
space, and then free the results of suio_flatten.  This is an
unnecessary malloc/free pair (which happened to trigger an
as-yet-unexplained out-of-memory error).  Instead, allocate
the memory as before but use copyout to write the marshalled
form directly into the pre-allocated memory.

Seems valgrind clean.