Add back-end support for object expiration.
Add an expiration time to the metadata database maintained by adbd.
The expiration time is passed in to adbd as an additional field for
ADBPROC_STORE.   The time is stored as a 32 bit unsigned integer,
interpreted to be the time since the Unix epoch; the value 0 is
a special case taken to mean that the data should not be expired.

Update the libadb API to allow users to set expiration times.
Use the updated API in dhblock servers, but for now preserving the
behavior of never expiring any data.

Expiration process can be triggered explicitly by new ADBPROC_EXPIRE
RPC or automatically when an insert detects that the system is at or
near capacity.  For the latter, the insertion process detects when
total capacity is at 90% (expire_threshold) of the quota and then
expires 24 (expire_batch_size) keys to make some more room.  Larger
batch sizes increase the time spent expiring and 24 seems empirically
not to slow adbd down by more than a second or so.  If expiration can
be made faster, perhaps we can run less frequently and in larger batches.

Bump up the number of locks and lock objects available acrosss the
board in dbfe_initialize_dbenv; expiration seems to cause a few more
of these to be used.