Add dhash.default_lifetime configuration parameter.
When set, use this value as the lifetime for any inserts to this node.
Eventually, it may make more sense to update the gateway protocol
to support expiration times, allowing the application to set desired
data lifetime.  However, this is easier to implement and is acceptable
for the UsenetDHT as it is currently deployed.