Add -l option to adbd to specify directory for BDB logs.
Using custom BDB configuration option to dbfe, allow user to specify
the location of log files on the command line.  No checking is
provided to help ensure that adbd is invoked with consistent arguments.
Beware.

Update start-dhash to use this when run with --logdir argument.

Refactor path canonicalization code to allow for use in both log and
data directories.