Reduce dbns cache size to 10MB from 30MB.
An overly large cache size may cause maintd problems when run in Carbonite
with even a few virtual nodes as it must mmap the cache for each node/ctype
combination in both adbd and maintd.  This seems to cause BDB to get an
ENOMEM and kill maintd with great unhappiness.

At the moment, having removed the datadb from adbd, a smaller cache
size works fine: current UsenetDHT deployment with 5 physical nodes
storing 5-6GB per vnode keeps the active set easily in about 6MB of
cache but we'll have to see how it works as the system scales up.