Do not check for no-repair conditions in lsd.
Goal is for maintd to control filtering of what repairs are generated
and leave for lsd the role of determining how quickly they are
processed.  When there are too few vnodes to make adequate repairs,
maintd already happens not to generate any repairs, but add a check
just in case.

maintd currently does not generate repairs in the single vnode case
for Carbonite because the getrepairs client in dhblock_*_srv asks for
repairs in the range [x, x].  This happens to be empty set.  Perhaps
it should really ask for repairs in the range [incID(x), x].  Passing
Tone happens to work here also because we do not bother syncing when
there are no other nodes in the system and PT only generates repairs
on sync.

Eventually maintd should be taught not to generate repairs that go to
other vnodes on the same physical node.  One way to do this would be
to have a single database for all virtual nodes: then the maintenance
code would see that they have all the same keys.