Fake child_hash when inconsistency detected in format_rpcnode.
Since the handle_send_node lookup and format_rpcnode calls are not
transaction protected, it is natural that occasionally there will
be an inconsistency found in the number of keys present.  Instead
of throwing a panic, simply attempt to fake out the process by
calculating a new hash using the first 64 keys returned.