Fix handling of range restriction in merkle_syncer.
In cases where merkle_syncer has to call getkeyrange to enumerate
remote keys, the limits to getkeyrange were being set incorrectly
when the range in question spanned 0 (i.e., max < min).  This resulted
in trees that were extremely truncated relative to the real keys present
on the source nodes and thus resulted in insufficient/spurious
repairs.

Add explicit handling for this case to enumerate the correct region using
piecewise clockwise sub-ranges.

Add test code to verify that the syncer works correctly in this case.