Request different repairs each time through in dhblock_srvs.
Old code was generating repairs always starting from the
predecessor's id.  This works fine for passingtone (which
ignores this start point) but not so well for carbonite:
maintd can not detect that a repair was completed until the
next sync round.  Therefore...

Add a last_repair variable to dhblock_srv subclasses to keep
track of the last repair supplied to us by maintd.  The next
time repairs are requested, start from there.  If no repairs
are returned, reset the pointer to the current predecessor.
Also, if the predecessor changes, jump up to the new predecesor.

Add a maint_pending variable to prevent multiple getrepairs
from coming back and confusing last_repair variable.

There is code duplication between dhblock_chash_srv and
dhblock_replicated_srv but it's not immediately obvious
how to factor it into the base class due to differences in
how the subclasses request/handle repairs.