Maintain copy of local tree for Carbonite repairs.
Carbonite calculates the number of available replicas of keys
based on the live local tree and local copies of the trees
of its successors.  In an environment where there are constant
writes (but by necessity, only periodic syncing), Carbonite will
calculate that only one replica exists for each key that has
been inserted into the local tree since the last sync with
its neighbors.  This would lead to repairs being spuriously
generated to its neighbors, simply because their copies have
not yet been detected.

To avoid these repairs, maintain a separate local copy of each
node's tree.  This limits the uncertainty to the amount of
fluctuation in the tree during the time it takes to make the copy.

With this fix, we now pass test_maint.sh for both carbonite and
passingtone.