Remove objects that have expired from the Merkle tree.
Objects that have expired do not need to be repaired; actually,
since not all nodes expire objects at the same rate, attempting to
repair such objects can lead to objects being repeatedly repaired
and then expired on a node that is at capacity.  This unnecessarily
chews up network, CPU and, most importantly, disk bandwidth.

To prevent this from happening, iterate over the expiration-ordered
index to identify and remove keys that have expired from the Merkle
tree.  This is done at start-up (before acknowledging the INITSPACE
RPC) and then periodically once a minute.  The data/metadata for
such objects remain in the database so they can be read until they
are expired when the quota is reached.

Implement by factoring out expiredb iteration code into expire_walk
so that it can be shared between regular data expiration and just
mtree expiration.  Generalize it to take a start and end time.
This isn't the cleanest in that expire_walk allocates memory that
it expects its callers to free.

expire_mtree uses an in-memory variable to incrementally expire
objects from the mtree instead of repeatedly starting from the oldest
objects.  Rather than write to this time to disk, we take the hit
of possibly repeating (and failing) expirations at startup.

This change means that we won't be able to rely on the Merkle tree
view of neighbors to help optimize reads (in the future), as there
may be objects stored on nodes that are not visible in the Merkle
tree.