Limit size of expiration transactions to 1000 keys.
I think that if we let too many keys be expired at once in
a transaction, we can run out of lock objects in the environment.
So, arbitrarily limit the size of these transactions to 1000 keys.
(We allow up to 5000 lock objects in utils/dbfe_sleepycat.C)

This might also help in the case where initial expiration takes
so long that appManager thinks that we failed to start and kills us
before restarting.  At least this way, we'll make progress toward
finishing.