Loosely synchronize when mtree cleaning is run.
Since we are cleaning out Merkle trees to remove expired objects,
we want to make sure everyone has approximately the same 'tip'.
This will help prevent cases where one node has just removed
an expired object but another one has not, leading to a repair
when the two nodes synchronize.

To avoid the potential of too many simultaneous cleanings on
a single node, perturb the synchronization just by a few seconds.

With this fix, it is not necessary to pre-expire objects from
the Merkle tree.  For now, I'm leaving in the code that leaves new
soon-to-expire insertions out of the tree though.  And, it is
very important that everyone has a loosely synchronized clock so
that objects expire at the same time!