Make expire time available to apps via dhashclient.
*** This changes the dhashgateway wire protocol! ***
If you have an application that uses its own dhashgateway implementation,
be sure to update in order to remain compatible.  The libdhashclient
interface remains the same but you will need to rebuild/relink in order
to get the updated wire protocol.

*** This changes the output format of dbm fetch! ***
If you have a script that parses the output of dbm fetch, please
check that it still parses the new output correctly.

Expose expiration as an option to dhashclient::insert (in an
option_block).  Users must set a new flag
DHASHCLIENT_EXPIRATION_SUPPLIED in order for the gateway to apply
the expiration.  Otherwise, the default lifetime will be used (as
configured on the gateway node).  Similarly, the expiration time of
the object is also passed back to the client in retrieve.  This is
implemented as a new expiration field in the insert/retrieve
arguments in dhashgateway_prot.x.  See tools/dbm.C for how this is
used.

Add a lifetime= option to dbm to allow the user to specify specific
object lifetimes during testing.  This can be set in units of
seconds, minutes, hours, days, weeks, months or years.