Handle PT repairs correctly for first and last successor.
PT talks to both the successor of the last replica and
the predecessor of the first replica.  If these nodes hold
the object, it is unlikely that a regular read would ever
consider those nodes.  Since the maintd synchronization actually
knows where the object came from, like for dhblock_replicated_srv,
we should take advantage of this and get the object from who
we heard about it from, without having to do a lookup that might
not even include that node.

Perhaps we should factor the src field into the parent
repair_job class at some point.