Initialize lsd parameters after creating chordnode.
Since maintd initializes its port based on the value in
parameters.addr.port, we need to make sure that it gets the
system-assigned port in the case where the user does not
specify a port with -p on the command line.  Therefore,
move the initialization to after the chordnode is set up
and grab the port via chordnode->get_port ().

Noted by "Diaz, Wilfredo (Freddy)" <wdiaz@alcatel-lucent.com>.