Share a single adbd DB over all virtual nodes.
Filter out vnodes from the same physical host when deciding
who to talk to for reads and writes (dhash/client.C).

Update Carbonite maintenance to only count copies of objects
on a given physical node once (maint/maint_policy.T).  We still
need to synchronize across all vnodes because we need to handle
all the particular ranges of responsibility correctly, so we
don't filter them out at the time we update the neighbor list
or perform the sync.

Update PT to sync with the first non-self neighbor.